base: compat: Move all implementations into removed_api.cpp
